While the goal is still to *not* optimize where it would undermine clarity, the cache misses appeared to be dominating the MatrixProduct extensions.
Reorganizing these extensions to use local references instead of explicit indexing in the inner loops reduced debug run times on the product of 2000x2000 matrices from 94 to 39 seconds (sequential) and 55 to 17 (parallel).

The loop pattern will feel further removed from the classical cij = sum(aik*bkj) fromula, but the intent is the same. And more importantly, the performance hit was too great to ignore.